backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
sentence-transformers==2.3.1
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.0
openai==1.12.0
pinecone-client==3.0.0